feat(agents-api): Set/get steps based on workflow state
