+++ agents-api/agents_api/clients/temporal.py
@@ -54,4 +54,5 @@ async def run_task_execution_workflow(
         task_queue=temporal_task_queue,
         id=str(job_id),
         run_timeout=timedelta(days=31),
+        # TODO: Should add search_attributes for queryability
     )
